DriverType renamed to Type and moved in DriverPropertyFactory
