+++ src/main/java/property_objects/DriverProperties.java
@@ -1,12 +1,13 @@
-package setups;
+package property_objects;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import setups.DriverPropertyFactory;
 
 @Data
 @NoArgsConstructor
 public class DriverProperties {
-    private DriverType type;
+    private DriverPropertyFactory.Type type;
     private boolean headless;
     private boolean incognito;
     private boolean maximize;
