--- src/main/java/setups/DriverType.java
@@ -1,16 +1,0 @@
-package setups;
-
-public enum DriverType {
-    CHROME("chrome"),
-    FIREFOX("firefox");
-
-    private final String value;
-
-    DriverType(String value) {
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
-}
