+++ src/main/java/utils/readers/DriverPropertyReader.java
@@ -1,28 +1,29 @@
-package setups;
+package utils.readers;
 
 import lombok.extern.log4j.Log4j2;
-import utils.PropertyReader;
+import property_objects.DriverProperties;
+import setups.DriverPropertyFactory;
 
 import java.io.IOException;
 
 @Log4j2
 public class DriverPropertyReader extends PropertyReader {
-    private static final String ACCOUNTS_PROPERTIES_FILE = "src/main/resources/driver.properties";
+    private static final String ACCOUNTS_PROPERTIES_FILE_PATH = "src/main/resources/driver.properties";
 
     public DriverPropertyReader() throws IOException {
-        super(ACCOUNTS_PROPERTIES_FILE);
+        super(ACCOUNTS_PROPERTIES_FILE_PATH);
     }
 
     public DriverProperties getDriverProperties() {
         DriverProperties driverProperties = new DriverProperties();
         String type = getString("type");
-        if (DriverType.CHROME.getValue().equals(type)) {
-            driverProperties.setType(DriverType.CHROME);
-        } else if (DriverType.FIREFOX.getValue().equals(type)) {
-            driverProperties.setType(DriverType.FIREFOX);
+        if (DriverPropertyFactory.Type.CHROME.getValue().equals(type)) {
+            driverProperties.setType(DriverPropertyFactory.Type.CHROME);
+        } else if (DriverPropertyFactory.Type.FIREFOX.getValue().equals(type)) {
+            driverProperties.setType(DriverPropertyFactory.Type.FIREFOX);
         } else {
             log.error("Unknown driverType '" + type + "', set ChromeDriver");
-            driverProperties.setType(DriverType.CHROME);
+            driverProperties.setType(DriverPropertyFactory.Type.CHROME);
         }
         driverProperties.setHeadless(getBoolean("headless"));
         driverProperties.setIncognito(getBoolean("incognito"));
